[webui] refactoring webui before implementing correct actix actors
